some small things are broken but overall its good
