+++ js/StdEnv.js
@@ -1,6 +1,10 @@
 import * as THREE from 'https://cdn.skypack.dev/pin/three@v0.135.0-pjGUcRG9Xt70OdXl97VF/mode=imports/optimized/three.js';
 import { PointerLockControls } from 'https://threejs.org/examples/jsm/controls/PointerLockControls.js';
 
+import { GLTFLoader } from 'https://threejs.org/examples/jsm/loaders/GLTFLoader.js';
+import * as BufferGeometryUtils from 'https://cdn.skypack.dev/three@0.135.0/examples/jsm/utils/BufferGeometryUtils.js';
+import { MeshBVH, MeshBVHVisualizer } from './util/three-mesh-bvh.js';
+
 import { DefaultDirectionalLight } from "./render/DefaultDirectionalLight.js"
 import { DefaultComposer } from "./render/DefaultComposer.js"
 import { Player } from './entities/Player.js';
@@ -12,7 +16,10 @@ const MEDIUM = 1;
 const HIGH = 2;
 const ULTRA = 3;
 
-let shadowLight;
+let shadowLight, collider;
+
+const bloomScene = new THREE.Scene();
+const clock = new THREE.Clock();
 
 /*
 This class wraps the whole Virtual Environement and provide an easy to use API for
@@ -43,11 +50,12 @@ Other usages should be like:
     for events on control lock and unlock
     VE.controls.addEventListener('lock', function() { .. do stuff });
     VE.controls.addEventListener('unlock', function() { .. do stuff });
+    VE.controls.lock();
 */
 
 class StdEnv {
     constructor() {  
-        this.init = function () {
+        this.init = function (terrainPath, x, y ,z) {
             this.graphicTier = localProxy.tier !== undefined ? localProxy.tier : 0;
 
             // ===== renderer =====
@@ -60,6 +68,7 @@ class StdEnv {
 
             // ===== scene =====
             this.scene = new THREE.Scene();
+            const scene = this.scene;
             this.scene.background = new THREE.Color(0x69e6f4);
             this.scene.fog = new THREE.Fog(0x69e6f4, 1600, 2000);
 
@@ -81,6 +90,60 @@ class StdEnv {
 
             // ===== shaders =====
             this.composer = new DefaultComposer(this.renderer, this.scene, this.camera);
+
+            // ===== load the terrain =====
+            function onProgress(xr) { console.log((xr.loaded / xr.total) * 100) }
+            function onError(e) { console.log(e) };
+
+            const loader = new GLTFLoader();
+            loader.load(terrainPath, function(object) {
+                object.scene.position.set(x, y, z);
+                object.scene.traverse(object => {
+                    if (object.isMesh) {
+                        object.castShadow = true;
+                        object.receiveShadow = true;
+                        object.material.roughness = 1;
+                        if (object.material.map) {
+                            object.material.map.anisotropy = 16;
+                            object.material.map.needsUpdate = true;
+                        }
+                        const cloned = object.clone();
+                        object.getWorldPosition(cloned.position);
+                        if (object.material.emissive && (object.material.emissive.r > 0 || object.material.emissive.g > 0 || object.material.color.b > 0)) {
+                            bloomScene.add(cloned);
+                        }
+                    }
+                    if (object.isLight) {
+                        object.parent.remove(object);
+                    }
+                });
+                scene.add(object.scene);
+                let geometries = [];
+                object.scene.traverse(object => {
+                    if (object.geometry && object.visible) {
+                        const cloned = object.geometry.clone();
+                        cloned.applyMatrix4(object.matrixWorld);
+                        for (const key in cloned.attributes) {
+                            if (key !== 'position') { cloned.deleteAttribute(key); }
+                        }
+                        geometries.push(cloned);
+                    }
+                });
+                const mergedGeometry = BufferGeometryUtils.mergeBufferGeometries(geometries, false);
+                mergedGeometry.boundsTree = new MeshBVH(mergedGeometry, { lazyGeneration: false });
+                collider = new THREE.Mesh(mergedGeometry);
+                collider.material.wireframe = true;
+                collider.material.opacity = 0.5;
+                collider.material.transparent = true;
+                collider.visible = false;
+                scene.add(collider);
+
+                const visualizer = new MeshBVHVisualizer(collider, 10);
+                visualizer.visible = false;
+                visualizer.update();
+
+                scene.add(visualizer);
+            }, onProgress, onError);
 
             // ===== player =====
             this.player = new Player();
@@ -139,6 +202,14 @@ class StdEnv {
 
 
     update(){
+        const delta = Math.min(clock.getDelta(), 0.1);
+        if (collider) {
+            for (let i = 0; i < 5; i++) {
+                this.player.update(delta / 5, this.camera, collider);
+                this.camera.position.copy(this.player.position);
+            }
+        }
+
         this.scene.fog.needsUpdate = true;
 
         if(this.graphicTier !== LOW){
